+++ packages/terra-form-validation/src/terra-dev-site/doc/example/FormSubmitDatePicker.jsx
@@ -44,7 +44,7 @@ export default class MainEntry extends React.Component {
         >
           {({ input, meta }) => (
             <TerraField
-              label="Enter your birthday"
+              label="Enter your planned start date"
               error={meta.error}
               isInvalid={meta.submitFailed && meta.error !== undefined}
               required