+++ packages/terra-disclosure-manager/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+4.20.0 - (September 6, 2019)
+------------------
 ### Changed
 * Cleaned up imports in examples and test files