+++ packages/terra-form-validation/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+1.8.0 - (September 6, 2019)
+------------------
 ### Changed
 * Cleaned up imports in examples and test files
 
@@ -17,6 +20,7 @@ Unreleased
 * updated package.json test scripts
 * Added `noValidate` attribute to all `<form />` elements in the examples
 * Added documentation on recommend usage of `noValidate` prop
+* Added DateInput form validation example
 
 1.5.0 - (July 30, 2019)
 ------------------